Add pre-SRoC licence agreement tests and tag supplementary billing specs
The licence agreement journey tests only covered the SRoC charge scheme, so there was no coverage confirming a pre-SRoC agreement flags a licence for the old charge scheme's supplementary bill run. New/end/delete specs now exist for both schemes, tagged @supplementary-billing (and @presroc for the pre-SRoC variants) so they can be run as a group.
